maylocnuoc/requirements.txt
Django
django-cors-headers==3.13.0
django-storages==1.13.2
Pillow==9.4.0
google-auth
google-cloud-storage
djangorestframework==3.14.0
django-tinymce==3.6.1